+++ R/get_germany_regional_cases.R
@@ -1,4 +1,6 @@
 #' Get German daily cases by Bundeslander
+#'
+#'
 #' @description Fetches COVID case counts by region in Germany.
 #' This data was collated by Jan-Philip Gehrcke (gh: jgehrcke) and is available at:
 #' https://github.com/jgehrcke/covid-19-germany-gae