+++ R/get_japan_regional_cases.R
@@ -1,8 +1,7 @@
-# Japan regional cases, daily
-# Source: https://en.wikipedia.org/wiki/2020_coronavirus_pandemic_in_Japan
-# NB. Not up to date: e.g. as of 23/3/20, latest data from 21/3/20
-# For an up to date map, see: https://mhlw-gis.maps.arcgis.com/apps/opsdashboard/index.html#/0c5d0502bbb54f9a8dddebca003631b8
-#
+#' Japan regional cases, daily
+
+#' @description Extracts regional case counts for Japan.
+#' [Source](https://en.wikipedia.org/wiki/2020_coronavirus_pandemic_in_Japan).
 #' @importFrom xml2 read_html
 #' @importFrom rvest html_nodes html_text html_table
 #' @importFrom tidyr as_tibble
@@ -10,7 +9,45 @@
 #' @importFrom stringr str_remove_all
 #' @importFrom lubridate as_date
 #' @importFrom memoise cache_filesystem memoise
-#
+#'
+#' @examples
+#'
+#'
+#' \dontrun{
+#'
+#' library(ggplot2)
+#' library(dplyr)
+#'
+#'  # Mapping: set up
+#'  data <- get_japan_regional_cases()
+#'
+#'  regions <- rnaturalearth::ne_states("Japan", returnclass = "sf")
+#'
+#'  regions_with_data <- regions %>%
+#'    dplyr::left_join(data, by = c("name" = "region")) %>%
+#'    dplyr::mutate(cases =  ifelse(is.na(cases), 0, cases)
+#'
+#'    # Map inset: Okinawa
+#'    jp_okinawa <- dplyr::filter(regions_with_data, name == "Okinawa")
+#'
+#'    jp_okinawa <- ggplot2::ggplot(jp_okinawa) +
+#'    geom_sf(aes(fill = cases)) +  coord_sf(datum = NA) +
+#'    xlab(jp_okinawa$region) +
+#'    theme_bw() +
+#'    theme(legend.position = "none")
+#'
+#'    ## Map: mainland with insets
+#'    jp_main <- dplyr::filter(regions_with_data, name != "Okinawa") %>%
+#'    ggplot() +
+#'    geom_sf(aes(fill = cases)) +
+#'    coord_sf(crs = sf::st_crs(4326), xlim = c(127, 146), ylim = c(29, 46)) +
+#'    theme_bw()
+#'
+#'    jp_main +
+#'    annotation_custom(
+#'    grob = ggplotGrob(jp_okinawa), xmin = 140,  xmax = 146,  ymin = 24,  ymax = 37)
+#'
+#' }
 
 get_japan_regional_cases <- function() {
   # Locate source
@@ -48,31 +85,4 @@ get_japan_regional_cases <- function() {
 
   return(cases)
 }
-
-
-# # Mapping: set up
-# data <- get_japan_regional_cases()
-# regions <- rnaturalearth::ne_states("Japan", returnclass = "sf")
-# regions_with_data <- regions %>%
-#   left_join(data, by = c("name" = "region")) %>%
-#   mutate(cases =  ifelse(is.na(cases), 0, cases))
-#
-# # Map inset: Okinawa
-# jp_okinawa <- dplyr::filter(regions_with_data, name == "Okinawa")
-# jp_okinawa <- ggplot2::ggplot(jp_okinawa) +
-#    geom_sf(aes(fill = cases)) +  coord_sf(datum = NA) +
-#    xlab(jp_okinawa$region) +
-#    theme_bw() +  theme(legend.position = "none")
-#
-# # Map: mainland with insets
-# jp_main <- dplyr::filter(regions_with_data, name != "Okinawa") %>%
-#   ggplot() +
-#     geom_sf(aes(fill = cases)) +
-#     coord_sf(crs = sf::st_crs(4326), xlim = c(127, 146), ylim = c(29, 46)) +
-#     theme_bw()
-#
-# jp_main +
-#   annotation_custom(
-#    grob = ggplotGrob(jp_okinawa), xmin = 140,  xmax = 146,  ymin = 24,  ymax = 37)
-